wikipedia/code/app/requirements.txt
python-dotenv

elasticsearch==8.*

sentence_transformers==2.2.2

flask==3.*
gunicorn
